--- .eslintrc
@@ -1,3 +1,0 @@
-{
-	"extends": [ "plugin:@wordpress/eslint-plugin/recommended" ]
-}
