--- .prettierrc
@@ -1,1 +1,0 @@
-"@wordpress/prettier-config"
