+++ package.json
@@ -8,9 +8,7 @@
     "start": "electron-forge start",
     "test": "echo \"Error: no test specified\" && exit 1",
     "package": "electron-forge package",
-    "make": "electron-forge make",
-    "lint": "eslint src/js/*",
-    "lint-fix": "eslint src/js/* --fix"
+    "make": "electron-forge make"
   },
   "author": "nagesh",
   "license": "MIT",
@@ -21,21 +19,11 @@
     "@electron-forge/maker-squirrel": "^6.0.0-beta.66",
     "@electron-forge/maker-zip": "^6.0.0-beta.66",
     "electron": "^20.2.0",
-    "eslint": "^7.30.0",
-    "eslint-config-prettier": "7.2.0",
-    "eslint-import-resolver-typescript": "2.4.0",
-    "eslint-plugin-jest": "24.3.6",
-    "eslint-plugin-jsdoc": "^39.3.14",
-    "eslint-plugin-jsdoc-alignment": "0.0.5",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-react": "7.22.0",
-    "eslint-plugin-react-hooks": "4.2.0",
     "tailwindcss": "^3.1.8"
   },
   "dependencies": {
     "dotenv": "^16.0.3",
     "electron-squirrel-startup": "^1.0.0",
-    "eslint-webpack-plugin": "^3.2.0",
     "mysql": "^2.18.1"
   },
   "config": {
@@ -67,5 +55,13 @@
       ]
     }
   },
-  "prettier": "@wordpress/prettier-config"
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/nagrampai/papyrus.git"
+  },
+  "keywords": [],
+  "bugs": {
+    "url": "https://github.com/nagrampai/papyrus/issues"
+  },
+  "homepage": "https://github.com/nagrampai/papyrus#readme"
 }
