+++ src/check-book.html
@@ -1,32 +1,62 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="stylesheet" href=".././dist/output.css">
-    <title>Papyrus @ Athena</title>
-  </head>
-  <body class="p-8">
+	<head>
+		<meta charset="UTF-8" />
+		<link rel="stylesheet" href=".././dist/output.css" />
+		<title>Papyrus @ Athena</title>
+	</head>
+	<body class="p-8">
+		<h1
+			class="text-4xl text-white w-11/12 m-auto bg-gray-700 text-center p-4 rounded-lg bg-opacity-50 font-extrabold mb-10"
+		>
+			Papyrus @ Athena
+		</h1>
+		<main class="w-11/12 m-auto">
+			<form
+				id="master-search-form"
+				action=""
+				target="_top"
+				class="text-center my-10"
+			>
+				<input
+					type="search"
+					name="book-or-member"
+					id="book-or-member"
+					placeholder="Book ID or Flat number"
+					required
+					class="py-2 px-4 border-2 w-3/12"
+				/>
+				<input
+					type="submit"
+					value="Search"
+					class="group w-1/12 justify-center rounded-md border border-transparent bg-indigo-600 py-2 px-3 text-sm font-medium text-white hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2 ml-5"
+				/>
+			</form>
+			<div id="result-container" class="grid grid-cols-12 mx-auto gap-4">
+				<div id="left-column" class="grid col-span-4 p-3"></div>
+				<div id="right-column" class="grid col-span-8"></div>
+			</div>
+		</main>
 
-    <h1 class="text-4xl text-white  w-11/12 m-auto bg-gray-700 text-center p-4 rounded-lg bg-opacity-50 font-extrabold mb-10">Papyrus @ Athena</h1>
-    <main class="w-11/12 m-auto">
-      <form id="master-search-form" action="" target="_top" class="text-center my-10">
-        <input type="search" name="book-or-member" id="book-or-member" placeholder="Book ID or Flat number" required class="py-2 px-4 border-2 w-3/12" >
-        <input type="submit" value="Search" class="group  w-1/12 justify-center rounded-md border border-transparent bg-indigo-600 py-2 px-3 text-sm font-medium text-white hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2 ml-5">
-      </form>
-      <div id="result-container" class="grid grid-cols-12 mx-auto gap-4">
-        <div id="left-column" class="grid col-span-4 p-3">
-        </div>
-        <div id="right-column" class="grid col-span-8">
-        </div>
-      </div>
-    </main>
+		<footer class="flex justify-center fixed bottom-5 w-full">
+			<a
+				href="./check-book.html"
+				class="text-blue-700 font-bold text-xl px-5"
+			>
+				Home </a
+			>|
+			<a
+				href="./member-master.html"
+				class="text-blue-700 font-bold text-xl px-5"
+				>Add Members</a
+			>|
+			<a
+				href="./add-book.html"
+				class="text-blue-700 font-bold text-xl px-5"
+				>Add Books</a
+			>
+		</footer>
 
-    <footer class="flex justify-center fixed bottom-5 w-full">
-      <a href="./check-book.html" class = "text-blue-700 font-bold text-xl px-5"> Home </a>|
-      <a href="./member-master.html" class = "text-blue-700 font-bold text-xl px-5">Add Members</a>|
-      <a href="./add-book.html" class = "text-blue-700 font-bold text-xl px-5">Add Books</a>
-    </footer>
-
-    <script type="module" src="./js/check-book.js"></script>
-  </body>
+		<script type="module" src="./js/check-book.js"></script>
+	</body>
 </html>
