+++ user-story.md
@@ -12,3 +12,4 @@
 
 ## Other notes / Constraints
 * Flat number should have an additional 0 for three digit flat numbers. Eg. D0703 , instead of D703.
+* Book ID cannot exceed 5 digits at the moment. This will need a minor code change to increase.
